chore: relocate test deploy helpers
