+++ lib/index.ts
@@ -1,7 +1,5 @@
 export * from "./account";
 export * from "./address";
-export * from "./aragon";
-export * from "./client";
 export * from "./constants";
 export * from "./contract";
 export * from "./deploy";
@@ -13,6 +11,7 @@ export * from "./ens";
 export * from "./event";
 export * from "./keccak";
 export * from "./log";
+export * from "./network";
 export * from "./oracle";
 export * from "./promise";
 export * from "./proxy";
@@ -23,4 +22,3 @@ export * from "./string";
 export * from "./time";
 export * from "./type";
 export * from "./units";
-export * from "./withdrawal";
