+++ test/0.4.24/lido/lido.staking.test.ts
@@ -6,7 +6,9 @@ import { HardhatEthersSigner } from "@nomicfoundation/hardhat-ethers/signers";
 
 import { ACL, Lido } from "typechain-types";
 
-import { certainAddress, deployLidoDao, ether, ONE_ETHER } from "lib";
+import { certainAddress, ether, ONE_ETHER } from "lib";
+
+import { deployAragonLidoDao } from "test/deploy";
 
 describe("Lido:staking", () => {
   let deployer: HardhatEthersSigner;
@@ -21,7 +23,7 @@ describe("Lido:staking", () => {
   beforeEach(async () => {
     [deployer, user] = await ethers.getSigners();
 
-    ({ lido, acl } = await deployLidoDao({ rootAccount: deployer, initialized: true }));
+    ({ lido, acl } = await deployAragonLidoDao({ rootAccount: deployer, initialized: true }));
 
     await acl.createPermission(user, lido, await lido.STAKING_CONTROL_ROLE(), deployer);
     await acl.createPermission(user, lido, await lido.STAKING_PAUSE_ROLE(), deployer);
