+++ test/deploy/index.ts
@@ -1,0 +1,2 @@
+export * from "./aragon";
+export * from "./withdrawal";
